requirements.txt
# Core
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Market Data & Trading
smartapi-python>=1.4.1
pandas>=2.1.0
numpy>=1.26.0

# Technical Analysis
ta>=0.11.0

# Database
sqlalchemy>=2.0.23

# Memory & RAG
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM & Agents
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langgraph>=0.0.30

# Data Lakehouse
pyiceberg>=0.6.0
minio>=7.2.0
pyarrow>=14.0.0
polars>=0.20.0

# Data Quality
great-expectations>=0.18.0

# Data Transforms
dbt-duckdb>=1.7.0
duckdb>=0.9.0

# API
fastapi>=0.108.0
uvicorn>=0.25.0

# Experiment Tracking
mlflow>=2.10.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
loguru>=0.7.2
httpx>=0.25.0
pyotp>=2.9.0
rich>=13.7.0

# Visualization (optional)
plotly>=5.18.0
streamlit>=1.29.0
